Set up createGrid() and deleteGrid()
Alongside these changes are a few style changes to the sketch pad, including making the grid pad itself square.
The createGrid() function utilizes flex-wrap and automatic calculation of width and height to make the pad functional.
the deleteGrid() function will be used later on when adding buttons to change the number of squares in the grid, as well as resetting any color changes to the grid.